+++ .github/workflows/release.yml
@@ -8,6 +8,7 @@ on:
       - '*-keycardai-starlette'
       - '*-keycardai-mcp'
       - '*-keycardai-mcp-fastmcp'
+      - '*-keycardai-fastmcp'
       - '*-keycardai-agents'
 
 jobs:
