+++ packages/fastmcp/examples/delegated_access/pyproject.toml
@@ -5,13 +5,13 @@ description = "GitHub API integration with Keycard delegated access using the @g
 readme = "README.md"
 requires-python = ">=3.10"
 dependencies = [
-    "keycardai-mcp-fastmcp",
+    "keycardai-fastmcp",
     "fastmcp>=3.0.0",
     "httpx>=0.27.0,<1.0.0",
 ]
 
 [tool.uv.sources]
-keycardai-mcp-fastmcp = { path = "../../", editable = true }
+keycardai-fastmcp = { path = "../../", editable = true }
 
 [project.scripts]
 delegated-access-server = "main:main"
