+++ pyproject.toml
@@ -61,6 +61,7 @@ exclude = []
 # Workspace sources will be defined here as packages are added
 keycardai-oauth = { workspace = true }
 keycardai-starlette = { workspace = true }
+keycardai-fastmcp = { workspace = true }
 keycardai-mcp-fastmcp = { workspace = true }
 keycardai-mcp = { workspace = true }
